requirements.txt
opencv-python>=4.9.0
onnxruntime>=1.18.0
numpy>=1.26.0
pandas>=2.2.0
openpyxl>=3.1.2
streamlit>=1.35.0
streamlit-webrtc>=0.47.0
av>=12.1.0
flask>=3.0.0
SQLAlchemy>=2.0.0
onvif-zeep>=0.2.12
wsdiscovery>=2.0.0
python-dotenv>=1.0.1
